Add HTML and CSS for structure and styling and JS code for fetching the quotes
This commit adds the HTML and CSS code necessary for styling the project.
It also implements two JS scripts, script.js for fetching the quotes from
an API (https://type.fit/api/quotes) and script-local.js for fetching
the quotes from a local file (quotes.js) so that the quote generator can
be used either way depending on the scripts specified in the HTML file.